workbench-c: Big pile of compiler de-linting

git-svn-id: https://svn.aros.org/svn/aros/trunk@37352 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/c/Version.c
@@ -1216,8 +1216,8 @@ int createdefvers(CONST_STRPTR name)
 	    parsedver.pv_vername &&
 	    parsedver.pv_revname)
 	{
-		sprintf(parsedver.pv_vername, "%ld", parsedver.pv_version);
-		sprintf(parsedver.pv_revname, ".%ld", parsedver.pv_revision);
+		sprintf(parsedver.pv_vername, "%ld", (long)parsedver.pv_version);
+		sprintf(parsedver.pv_revname, ".%ld", (long)parsedver.pv_revision);
 
 		return RETURN_OK;
 	}
@@ -1591,7 +1591,7 @@ int setvervar(CONST_STRPTR name, LONG ver, LONG rev)
 {
 	UBYTE buf[32];
 
-	sprintf(buf, "%ld.%ld", (LONG) ver, (LONG) rev);
+	sprintf(buf, "%ld.%ld", (long) ver, (long) rev);
 
 	return SetVar((STRPTR) name, buf, -1, GVF_LOCAL_ONLY | LV_VAR) ? RETURN_OK : -1;
 }
